fix preprocess.py

+++ AI/src/preprocessing/VideoPreprocessor.py
@@ -51,6 +51,10 @@ class VideoPreprocessor(object):
     def spath(self) -> str:
         return self.__spath
 
+    @property
+    def is_label(self) -> bool:
+        return self.__is_labeled
+
     def _is_labeled(self, ds_name: str) -> bool:
         flag: bool = False
         path_components: List[str] = self.__fpath.split(os.sep)
@@ -110,46 +114,48 @@ class VideoPreprocessor(object):
             c/ Save video stream as .pt file
          and
         """
-        video: torch.Tensor = v2(self.__spath, device=self.__device)  # [T,H,W,C] in cpu device
-
-        total_frames: int = video.shape[0]
-        seg_start_idx: torch.Tensor = torch.linspace(0, total_frames, self.__num_segments).clamp(0, total_frames).int()
-
-        save_tensor: None | torch.Tensor = None
-        if not self.__is_labeled:
-            for i in range(0, len(seg_start_idx)-1):
-                start, end = seg_start_idx[i].item(), seg_start_idx[i + 1].item()
-
-                indices: torch.Tensor = torch.arange(start, end, device=video.device)
-                inter_mode = "nearest-exact" if indices.shape[0] > self.__num_frames else "trilinear"
-
-                frames: torch.Tensor = torch.index_select(video, 0, indices)
-
-                try:
-                    frames = frames.to(self.__device).permute(-1, 0, 1, 2).unsqueeze(0)
-                except torch.cuda.OutOfMemoryError:
-                    frames = frames.permute(-1, 0, 1, 2).unsqueeze(0)
-
-                frames = torch.nn.functional.interpolate(
-                    frames.type(torch.float32) if inter_mode == "trilinear" else frames,
-                    (self.__num_frames, *frames.shape[-2:]),
-                    mode=inter_mode
-                    )
-                frames = frames.to("cpu").type(torch.uint8)
-
-                save_tensor = frames if save_tensor is None else torch.vstack((save_tensor, frames))
-        else:
-            save_tensor = video
-
         extension: str = pathlib.Path(self.spath).name.split(".")[-1]
-        torch.save(save_tensor, self.spath.replace(extension, "pt"))
-        torch.serialization.add_safe_globals([save_tensor])
+        if not os.path.isfile(self.spath.replace(extension, "pt")):
+            video: torch.Tensor = v2(self.__spath, device=self.__device)  # [T,H,W,C] in cpu device
 
-        if del_prev_result:
-            os.remove(self.__spath)
+            total_frames: int = video.shape[0]
+            seg_start_idx: torch.Tensor = torch.linspace(0, total_frames, self.__num_segments).clamp(0, total_frames).int()
 
-        if self.__device == "cuda":
-            del video
-            gc.collect()
-            torch.cuda.empty_cache()
+            save_tensor: None | torch.Tensor = None
+            if not self.__is_labeled:
+                for i in range(0, len(seg_start_idx)-1):
+                    start, end = seg_start_idx[i].item(), seg_start_idx[i + 1].item()
+
+                    indices: torch.Tensor = torch.arange(start, end, device=video.device)
+                    inter_mode = "nearest-exact" if indices.shape[0] > self.__num_frames else "trilinear"
+
+                    frames: torch.Tensor = torch.index_select(video, 0, indices)
+
+                    try:
+                        frames = frames.to(self.__device).permute(-1, 0, 1, 2).unsqueeze(0)
+                    except torch.cuda.OutOfMemoryError:
+                        frames = frames.permute(-1, 0, 1, 2).unsqueeze(0)
+
+                    frames = torch.nn.functional.interpolate(
+                        frames.type(torch.float32) if inter_mode == "trilinear" else frames,
+                        (self.__num_frames, *frames.shape[-2:]),
+                        mode=inter_mode
+                        )
+                    frames = frames.to("cpu").type(torch.uint8)
+
+                    save_tensor = frames if save_tensor is None else torch.vstack((save_tensor, frames))
+            else:
+                save_tensor = video
+
+            extension: str = pathlib.Path(self.spath).name.split(".")[-1]
+            torch.save(save_tensor, self.spath.replace(extension, "pt"))
+            torch.serialization.add_safe_globals([save_tensor])
+
+            if del_prev_result:
+                os.remove(self.__spath)
+
+            if self.__device == "cuda":
+                del video
+                gc.collect()
+                torch.cuda.empty_cache()
         return None
